chore: move action components to page dir
