+++ src/pages/actions/Activation.js
@@ -1,18 +1,18 @@
-import NavBackHome from './NavBackHome';
-import NavBar from './NavBar';
+import NavBackHome from '../../components/common/NavBackHome';
+import NavBar from '../../components/common/NavBar';
 import { userSessionState } from '../../lib/auth';
 import { useAtom } from 'jotai';
-import Unauthorized from './Unauthorized';
+import Unauthorized from '../../components/common/Unauthorized';
 
-export default function CityTools(props) {
+export default function CityActivation(props) {
   const [userSession] = useAtom(userSessionState);
   if (userSession.isUserSignedIn()) {
     return (
       <>
         <NavBar city={props.config.cityName} symbol={props.token.symbol} path={props.path} />
-        <h3>{props.token.symbol} Tools</h3>
-        <p>Get User ID for an address</p>
-        <p>Other tools that don't fit</p>
+        <h3>{props.token.symbol} Activation</h3>
+        <p>Registration action</p>
+        <p>If registered, show registration info</p>
         <hr />
         <NavBackHome />
       </>
