+++ src/pages/actions/Dashboard.js
@@ -1,8 +1,8 @@
-import MiningActivity from '../dashboard/MiningActivity';
-import StackingActivity from '../dashboard/StackingActivity';
-import TransactionLog from '../dashboard/TransactionLog';
-import NavBackHome from './NavBackHome';
-import NavBar from './NavBar';
+import MiningActivity from '../../components/dashboard/MiningActivity';
+import StackingActivity from '../../components/dashboard/StackingActivity';
+import TransactionLog from '../../components/dashboard/TransactionLog';
+import NavBackHome from '../../components/common/NavBackHome';
+import NavBar from '../../components/common/NavBar';
 
 export default function CityDashboard(props) {
   return (
