+++ src/pages/actions/Stacking.js
@@ -1,18 +1,17 @@
-import NavBackHome from './NavBackHome';
-import NavBar from './NavBar';
+import StackingContainer from '../../components/stacking/StackingContainer';
+import NavBackHome from '../../components/common/NavBackHome';
+import NavBar from '../../components/common/NavBar';
 import { userSessionState } from '../../lib/auth';
 import { useAtom } from 'jotai';
-import Unauthorized from './Unauthorized';
+import Unauthorized from '../../components/common/Unauthorized';
 
-export default function CityActivation(props) {
+export default function CityStacking(props) {
   const [userSession] = useAtom(userSessionState);
   if (userSession.isUserSignedIn()) {
     return (
       <>
         <NavBar city={props.config.cityName} symbol={props.token.symbol} path={props.path} />
-        <h3>{props.token.symbol} Activation</h3>
-        <p>Registration action</p>
-        <p>If registered, show registration info</p>
+        <StackingContainer contracts={props.contracts} token={props.token} config={props.config} />
         <hr />
         <NavBackHome />
       </>
