+++ src/pages/actions/Tools.js
@@ -1,17 +1,18 @@
-import StackingContainer from '../stacking/StackingContainer';
-import NavBackHome from './NavBackHome';
-import NavBar from './NavBar';
+import NavBackHome from '../../components/common/NavBackHome';
+import NavBar from '../../components/common/NavBar';
 import { userSessionState } from '../../lib/auth';
 import { useAtom } from 'jotai';
-import Unauthorized from './Unauthorized';
+import Unauthorized from '../../components/common/Unauthorized';
 
-export default function CityStacking(props) {
+export default function CityTools(props) {
   const [userSession] = useAtom(userSessionState);
   if (userSession.isUserSignedIn()) {
     return (
       <>
         <NavBar city={props.config.cityName} symbol={props.token.symbol} path={props.path} />
-        <StackingContainer contracts={props.contracts} token={props.token} config={props.config} />
+        <h3>{props.token.symbol} Tools</h3>
+        <p>Get User ID for an address</p>
+        <p>Other tools that don't fit</p>
         <hr />
         <NavBackHome />
       </>
