+++ src/pages/cities/Miami.js
@@ -3,12 +3,12 @@ import { Redirect, Router } from '@reach/router';
 import { useAtom } from 'jotai';
 import { currentCity, currentCitySymbol } from '../../store/common';
 import { MIA_CONTRACTS, MIA_TOKEN, MIA_CONFIG } from '../../store/miami';
-import CityDashboard from '../../components/common/CityDashboard';
-import CityStats from '../../components/common/CityStats';
-import CityActivation from '../../components/common/CityActivation';
-import CityMining from '../../components/common/CityMining';
-import CityStacking from '../../components/common/CityStacking';
-import CityTools from '../../components/common/CityTools';
+import Dashboard from '../../pages/actions/Dashboard';
+import Stats from '../../pages/actions/Stats';
+import Activation from '../../pages/actions/Activation';
+import Mining from '../../pages/actions/Mining';
+import Stacking from '../../pages/actions/Stacking';
+import Tools from '../../pages/actions/Tools';
 import NotFound from '../NotFound';
 
 export default function Miami() {
@@ -26,27 +26,17 @@ export default function Miami() {
   return (
     <Router>
       <Redirect from="/" to="dashboard" />
-      <CityDashboard
-        path="dashboard"
-        contracts={MIA_CONTRACTS}
-        token={MIA_TOKEN}
-        config={MIA_CONFIG}
-      />
-      <CityStats path="stats" contracts={MIA_CONTRACTS} token={MIA_TOKEN} config={MIA_CONFIG} />
-      <CityActivation
+      <Dashboard path="dashboard" contracts={MIA_CONTRACTS} token={MIA_TOKEN} config={MIA_CONFIG} />
+      <Stats path="stats" contracts={MIA_CONTRACTS} token={MIA_TOKEN} config={MIA_CONFIG} />
+      <Activation
         path="activation"
         contracts={MIA_CONTRACTS}
         token={MIA_TOKEN}
         config={MIA_CONFIG}
       />
-      <CityMining path="mining" contracts={MIA_CONTRACTS} token={MIA_TOKEN} config={MIA_CONFIG} />
-      <CityStacking
-        path="stacking"
-        contracts={MIA_CONTRACTS}
-        token={MIA_TOKEN}
-        config={MIA_CONFIG}
-      />
-      <CityTools path="tools" contracts={MIA_CONTRACTS} token={MIA_TOKEN} config={MIA_CONFIG} />
+      <Mining path="mining" contracts={MIA_CONTRACTS} token={MIA_TOKEN} config={MIA_CONFIG} />
+      <Stacking path="stacking" contracts={MIA_CONTRACTS} token={MIA_TOKEN} config={MIA_CONFIG} />
+      <Tools path="tools" contracts={MIA_CONTRACTS} token={MIA_TOKEN} config={MIA_CONFIG} />
       <NotFound default />
     </Router>
   );
